update stock and verification on EndDay and download

+++ app/src/main/java/com/optimus/eds/ui/home/HomeRepository.java
@@ -14,6 +14,7 @@ import com.optimus.eds.db.dao.OrderDao;
 import com.optimus.eds.db.dao.ProductsDao;
 import com.optimus.eds.db.dao.RouteDao;
 
+import com.optimus.eds.model.BaseResponse;
 import com.optimus.eds.model.PackageProductResponseModel;
 import com.optimus.eds.model.RouteOutletResponseModel;
 import com.optimus.eds.source.API;
@@ -22,8 +23,13 @@ import com.optimus.eds.utils.PreferenceUtil;
 
 
 import java.io.IOException;
+import java.util.Calendar;
+import java.util.HashMap;
 import java.util.concurrent.Executor;
 
+import io.reactivex.SingleObserver;
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.disposables.Disposable;
 import io.reactivex.observers.DisposableSingleObserver;
 import io.reactivex.schedulers.Schedulers;
 import retrofit2.Response;
@@ -41,6 +47,8 @@ public class HomeRepository {
 
 
     private MutableLiveData<Boolean> isLoading;
+    private MutableLiveData<String> msg;
+    private MutableLiveData<Boolean> onDayStartLiveData;
     private API webService;
     private Executor executor;
 
@@ -59,6 +67,8 @@ public class HomeRepository {
         merchandiseDao = appDatabase.merchandiseDao();
         customerDao = appDatabase.customerDao();
         isLoading = new MutableLiveData<>();
+        onDayStartLiveData = new MutableLiveData<>();
+        msg = new MutableLiveData<>();
         webService = api;
         this.executor = executor;
 
@@ -70,29 +80,32 @@ public class HomeRepository {
             @Override
             public void onSuccess(TokenResponse tokenResponse) {
                 preferenceUtil.saveToken(tokenResponse.getAccessToken());
-                fetchTodayData();
+                fetchTodayData(true);
             }
 
             @Override
             public void onError(Throwable e) {
                 isLoading.postValue(false);
+                msg.postValue(e.getMessage());
             }
         });
 
     }
 
-    public void fetchTodayData(){
+    public void fetchTodayData(boolean onDayStart){
         //@TODO this needs to be done
         // boolean ifDataAlreadyExist = dao.hasRefreshedData();
         executor.execute(() -> {
             try {
                 Response<RouteOutletResponseModel> response = webService.loadTodayRouteOutlets().execute();
                 if(response.isSuccessful()){
-                    routeDao.deleteAllRoutes();
                     routeDao.deleteAllOutlets();
+                    routeDao.deleteAllRoutes();
                     routeDao.deleteAllAssets();
-                    merchandiseDao.deleteAllMerchandise();
-                    customerDao.deleteAllCustomerInput();
+                    if(onDayStart) {
+                        merchandiseDao.deleteAllMerchandise();
+                        customerDao.deleteAllCustomerInput();
+                    }
                     routeDao.insertRoutes(response.body().getRouteList());
                     routeDao.insertOutlets(response.body().getOutletList());
                     routeDao.insertAssets(response.body().getAssetList());
@@ -116,7 +129,10 @@ public class HomeRepository {
             try {
                 Response<PackageProductResponseModel> response = webService.loadTodayPackageProduct().execute();
                 if(response.isSuccessful()){
-                    orderDao.deleteAllOrders();
+                    if(onDayStart) {
+                        orderDao.deleteAllOrders();
+                        updateWorkStatus(onDayStart);
+                    }
 
                     productsDao.deleteAllPackages();
                     productsDao.deleteAllProductGroups();
@@ -124,15 +140,17 @@ public class HomeRepository {
                     productsDao.insertProductGroups(response.body().getProductGroups());
                     productsDao.insertPackages(response.body().getPackageList());
                     productsDao.insertProducts(response.body().getProductList());
+
                 }
                 else{
-
+                msg.postValue(response.errorBody().string());
                 }
             } catch (IOException e) {
                 e.printStackTrace();
                 Log.e(TAG,e.getMessage());
             }finally {
                 isLoading.postValue(false);
+
             }
 
 
@@ -141,7 +159,42 @@ public class HomeRepository {
 
     }
 
+    public void updateWorkStatus(boolean isStart){
+        HashMap<String, Long> map = new HashMap<>();
+        map.put(isStart?"startDay":"endDay", Calendar.getInstance().getTimeInMillis());
+        webService.updateStartEndStatus(map).observeOn(AndroidSchedulers.mainThread())
+                .subscribeOn(Schedulers.io()).subscribe(new SingleObserver<BaseResponse>() {
+            @Override
+            public void onSubscribe(Disposable d) {
+
+            }
+
+            @Override
+            public void onSuccess(BaseResponse baseResponse) {
+                isLoading.postValue(false);
+                if(baseResponse.isSuccess()){
+                    onDayStartLiveData.postValue(isStart);
+                }else
+                    msg.postValue(baseResponse.getResponseMsg());
+            }
+
+            @Override
+            public void onError(Throwable e) {
+                isLoading.postValue(false);
+            msg.postValue(e.getMessage());
+            }
+        });
+    }
+
     public LiveData<Boolean> mLoading() {
         return isLoading;
+    }
+
+    public LiveData<Boolean> startDay(){
+        return onDayStartLiveData;
+    }
+
+    public LiveData<String> getError() {
+        return msg;
     }
 }
